Stl boundary:
- terrain contact optimization
- make file optimization for linux compilation
- removing all warnings for linux and linux compilation

+++ src/DynamicRelaxation.cpp
@@ -22,7 +22,7 @@ double DynamicRelaxation::computeKineticEnergy(vector<Body*>* bodies)
         
         // for each particle
         #pragma omp parallel for reduction(+:currentKineticEnergy) shared(particles)
-        for (int i = 0; i < particles->size(); ++i) {
+        for (int i = 0; i < static_cast<int> (particles->size()); ++i) {
 
             // verify active particle
             if (!particles->at(i)->getActive()) { continue; }
@@ -71,7 +71,7 @@ void DynamicRelaxation::setStaticSolution(vector<Body*>* bodies, int loopCounter
             
             // for each particle
             #pragma omp parallel for shared(particles)
-            for (int i = 0; i < particles->size(); ++i) {
+            for (int i = 0; i < static_cast<int> (particles->size()); ++i) {
 
                 // verify active particle
                 if (!particles->at(i)->getActive()) { continue; }
